pyproject.toml
[tool.poetry]
name = "prostate-mri-classification"
version = "0.1.0"
description = ""
authors = ["Kevin Mathew T"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
h5py = "^3.12.1"
numpy = "^1.24"
pandas = "^2.2.3"
torch = "^2.5.1"
scikit-learn = "^1.6.0"
pyyaml = "^6.0.2"
torchvision = "^0.20.1"
accelerate = "^1.2.1"
wandb = "^0.19.1"
notebook = "^7.3.2"
monai = "^1.4.0"
einops = "^0.8.0"
torchio = "^0.20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
